+++ migrations/2023-09-16-201255_moderation_kicks/up.sql
@@ -1,8 +1,8 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS "moderation_kicks" (
-	"kick_id"	INTEGER,
-	"target_id"	INTEGER NOT NULL,
-	"moderator_id"	INTEGER NOT NULL,
+	"kick_id"	BIGINT,
+	"target_id"	BIGINT NOT NULL,
+	"moderator_id"	BIGINT NOT NULL,
 	"kick_reason"	TEXT,
 	FOREIGN KEY("target_id") REFERENCES "core_users"("user_id"),
 	FOREIGN KEY("moderator_id") REFERENCES "core_users"("user_id"),
